GDAL

+++ gView.DataSources.GDAL/DataSources/GDAL/RasterClass.cs
@@ -377,7 +377,7 @@ namespace gView.DataSources.GDAL
                                 break;
                         }
                         band.ReadRaster(x, y, wWidth, wHeight,
-                            new IntPtr(buf.ToInt32() + ch),
+                            new IntPtr(buf.ToInt64() + ch),
                             iWidth, iHeight, OSGeo.GDAL.DataType.GDT_Byte, pixelSpace, stride);
 
                         band.Dispose();
@@ -448,7 +448,7 @@ namespace gView.DataSources.GDAL
                                 break;
                         }
                         band.ReadRaster(x, y, wWidth, wHeight,
-                                new IntPtr(buf.ToInt32() + ch),
+                                new IntPtr(buf.ToInt64() + ch),
                                 iWidth, iHeight, OSGeo.GDAL.DataType.GDT_Byte, pixelSpace, stride);
 
                         band.Dispose();
